pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-aqi-control-env"
version = "0.1.0"
description = "City-level AQI policy controller benchmark for OpenEnv — Indian NCAP cities"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
inference = [
    "openai>=1.0.0",
    "requests>=2.25.0",
]

[project.scripts]
server = "aqi_control_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["aqi_control_env", "aqi_control_env.env", "aqi_control_env.server"]
package-dir = { "aqi_control_env" = ".", "aqi_control_env.env" = "env", "aqi_control_env.server" = "server" }

[tool.setuptools.package-data]
"aqi_control_env" = ["data/**/*.json", "openenv.yaml"]
